перенесет render, createElement в отдельный модуль, создаст функции для работы с компонентами replace, remove
