+++ src/utils/utils.js
@@ -1,5 +1,3 @@
-import {RenderPosition} from "./const";
-
 const castTimeFormat = (value) => {
   return value < 10 ? `0${value}` : String(value);
 };
@@ -11,22 +9,4 @@ const createTimeFormat = (date) => {
   return `${hours}:${minutes}`;
 };
 
-const createElement = (template) => {
-  const element = document.createElement(`div`);
-  element.innerHTML = template;
-
-  return element.firstChild;
-};
-
-const render = (container, component, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(component.getElement());
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(component.getElement());
-      break;
-  }
-};
-
-export {createTimeFormat, createElement, render};
+export {createTimeFormat};
